requirements.txt
pandas==2.0.0
psycopg2==2.9.3
python-dotenv==0.21.0
numpy==1.24.0
requests==2.28.1
apache-airflow==2.6.0

